reviving scrollToTop behavior
